use sum values from store
